requirements.txt
alabaster==0.7.10
anaconda-client==1.6.3
anaconda-navigator==1.6.2
anaconda-project==0.6.0
appnope==0.1.0
appscript==1.0.1
argcomplete==1.0.0
asgi-redis==0.8.3
asgiref==0.9
asn1crypto==0.22.0
astroid==1.4.9
astropy==1.3.2
autobahn==0.12.1
Babel==2.4.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bitarray==0.8.1
blaze==0.10.1
bleach==1.5.0
bokeh==0.12.5
boto==2.46.1
boto3==1.4.6
botocore==1.6.8
Bottleneck==1.2.1
certifi==2018.4.16
cffi==1.10.0
channels==0.9.3
chardet==3.0.3
chest==0.2.3
click==6.7
cloudpickle==0.2.2
clyent==1.2.2
colorama==0.3.9
conda==4.3.30
conda-build==3.0.19
conda-verify==2.0.0
configobj==5.0.6
constantly==15.1.0
contextlib2==0.5.5
cryptography==1.8.1
cycler==0.10.0
Cython==0.25.2
cytoolz==0.8.2
daphne==0.9.1
dask==0.14.3
datashape==0.5.4
decorator==4.0.11
dill==0.2.6
distributed==1.16.3
dj-database-url==0.5.0
Django==2.2.2
django-bootstrap3==8.2.1
django-bootstrap4==0.0.7
django-heroku==0.3.1
django-js-asset==1.1.0
django-mptt==0.9.1
djangorestframework==3.9.4
docutils==0.13.1
dynd==0.7.3.dev1
entrypoints==0.2.2
enum34==1.1.6
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==2.0.7
flake8==3.3.0
Flask==0.12.2
Flask-Cors==3.0.2
gevent==1.2.1
glob2==0.5
greenlet==0.4.12
gunicorn==19.4.5
h5py==2.7.0
haikunator==1.0.1
HeapDict==1.0.0
html5lib==0.9999999
idna==2.5
imagesize==0.7.1
incremental==16.10.1
ipykernel==4.6.1
ipython==5.3.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
isort==4.2.5
itsdangerous==0.24
jdcal==1.3
jedi==0.10.2
Jinja2==2.9.6
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
lazy-object-proxy==1.2.2
llvmlite==0.18.0
locket==0.2.0
lxml==3.7.3
Markdown==2.6.11
MarkupSafe==0.23
matplotlib==2.0.2
mccabe==0.6.1
mistune==0.7.4
mpi4py==2.0.0
mpmath==0.19
msgpack-python==0.4.8
multipledispatch==0.4.9
navigator-updater==0.1.0
nb-anacondacloud==1.4.0
nb-conda==2.2.0
nb-conda-kernels==2.1.0
nbconvert==5.1.1
nbformat==4.3.0
nbpresent==3.0.2
networkx==1.11
nltk==3.2.3
nose==1.3.7
notebook==5.0.0
numba==0.33.0
numexpr==2.6.2
numpy==1.14.0
numpydoc==0.6.0
odo==0.5.0
olefile==0.44
openpyxl==2.4.7
packaging==16.8
pandas==0.20.1
pandocfilters==1.4.1
partd==0.3.8
pathlib2==2.2.1
patsy==0.4.1
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
pkginfo==1.4.1
ply==3.10
postgres==2.2.1
prompt-toolkit==1.0.14
protobuf==3.5.1
psutil==5.2.2
psycopg2==2.7.1
ptyprocess==0.5.1
py==1.4.33
pyasn1==0.2.3
pycodestyle==2.3.1
pycosat==0.6.2
pycparser==2.17
pycrypto==2.6.1
pycurl==7.43.0
pyflakes==1.5.0
Pygments==2.2.0
pylint==1.6.4
pyodbc==4.0.16
pyOpenSSL==17.0.0
pyparsing==2.1.4
pytest==3.0.7
python-dateutil==2.6.0
python-mnist==0.3
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.0
QtPy==1.2.1
redis==2.10.5
requests==2.14.2
rope-py3k==0.9.4.post1
s3transfer==0.1.13
scikit-image==0.13.0
scikit-learn==0.18.1
scipy==1.0.0
seaborn==0.7.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
snowballstemmer==1.2.1
sockjs-tornado==1.0.3
sortedcollections==0.5.3
sortedcontainers==1.5.7
Sphinx==1.5.6
sphinx-rtd-theme==0.2.4
spyder==3.1.4
SQLAlchemy==1.1.9
sqlparse==0.3.0
statsmodels==0.8.0
sympy==1.0
tables==3.3.0
tblib==1.3.2
tensorflow-tensorboard==0.4.0rc3
terminado==0.6
testpath==0.3
Theano==0.8.2
toolz==0.8.2
tornado==4.5.1
traitlets==4.3.2
Twisted==15.5.0
txaio==2.2.1
unicodecsv==0.14.1
virtualenv==15.1.0
wcwidth==0.1.7
Werkzeug==0.14.1
whitenoise==2.0.6
widgetsnbextension==2.0.0
wrapt==1.10.10
xgboost==0.6
xlrd==1.0.0
XlsxWriter==0.9.6
xlwings==0.10.4
xlwt==1.2.0
zict==0.1.2
zope.interface==4.1.3
